# specs/001-hugo-matter-cms/contracts/api-openapi.yaml
openapi: 3.0.3
info:
  title: Hugo Matters API
  version: 0.1.0
  description: |
    Local ApiService contract for Hugo Matters CMS v1.
    Web (Blazor Server) is the primary client via Aspire service discovery.
servers:
  - url: https://apiservice
    description: Aspire service discovery name (scheme may resolve to http in local dev)
...
paths:
...
  /api/connection:
    get:
      tags: [Connection]
      summary: Get current connected site (v1 single site)
      responses:
        "200":
          description: Connected site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectedSite"
        "404":
          description: No site connected
    post:
      tags: [Connection]
      summary: Bind a Hugo site repository using a GitHub App installation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectRequest"
      responses:
        "200":
          description: Connected site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectedSite"
        "400":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Connection]
      summary: Disconnect local site binding (does not uninstall GitHub App)
      responses:
        "204":
          description: Disconnected
...
  /api/content:
    get:
      tags: [Content]
      summary: List posts/pages visible in the active session tree
      parameters:
        - name: contentType
          in: query
          schema:
            type: string
            enum: [post, page, all]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentItemSummary"
        "404":
          description: No active session
    post:
      tags: [Content]
      summary: Create new post/page in session buffer using theme-pack defaults
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentCreateRequest"
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
        "400":
          $ref: "#/components/responses/Error"
  /api/content/{path}:
    get:
      tags: [Content]
      summary: Get a content item by repo-relative path
      parameters:
        - $ref: "#/components/parameters/ContentPath"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    put:
      tags: [Content]
      summary: Upsert content in the local session buffer (not Git until save)
      parameters:
        - $ref: "#/components/parameters/ContentPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentItemWrite"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
        "400":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Content]
      summary: Mark content deleted in session buffer (persisted on save)
      parameters:
        - $ref: "#/components/parameters/ContentPath"
      responses:
        "204":
          description: Marked deleted
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
  /api/preview/editor:
    post:
      tags: [Preview]
      summary: Approximate in-editor HTML preview (shortcodes may not render)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditorPreviewRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditorPreviewResponse"
  /api/preview/site:
    post:
      tags: [Preview]
      summary: Start real Hugo site preview from saved session branch tip
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SitePreview"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          description: Preview already running or session missing
    get:
      tags: [Preview]
      summary: Get current site preview status
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SitePreview"
        "404":
          description: No preview
    delete:
      tags: [Preview]
      summary: Stop site preview and clean workspace
      responses:
        "204":
          description: Stopped
...
components:
  parameters:
    ContentPath:
      name: path
      in: path
      required: true
      description: Repo-relative path (URL-encoded)
      schema:
        type: string
  responses:
    Error:
      description: Safe failure payload (no secrets)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    ConnectRequest:
      type: object
      required: [installationId, owner, repo]
      properties:
        installationId:
          type: integer
          format: int64
        owner:
          type: string
        repo:
          type: string
    ConnectedSite:
      type: object
      required: [id, ownerLogin, repoName, defaultBranch, themePackId, status]
      properties:
        id:
          type: string
          format: uuid
        ownerLogin:
          type: string
        repoName:
          type: string
        defaultBranch:
          type: string
        htmlUrl:
          type: string
        themePackId:
          type: string
        themePackVersion:
          type: string
        status:
          type: string
          enum: [Connected, AccessLost, Disconnected]
...
    ContentItemSummary:
      type: object
      required: [path, contentType, title]
      properties:
        path:
          type: string
        contentType:
          type: string
          enum: [post, page]
        title:
          type: string
        isDeleted:
          type: boolean
        isNew:
          type: boolean
    ContentItem:
      type: object
      required: [path, contentType, frontMatter, body]
      properties:
        path:
          type: string
        contentType:
          type: string
          enum: [post, page]
        frontMatter:
          type: object
          additionalProperties: true
        body:
          type: string
        isDeleted:
          type: boolean
        isNew:
          type: boolean
        hasUnsavedLocalEdits:
          type: boolean
    ContentItemWrite:
      type: object
      required: [frontMatter, body]
      properties:
        frontMatter:
          type: object
          additionalProperties: true
        body:
          type: string
    ContentCreateRequest:
      type: object
      required: [contentType, slug]
      properties:
        contentType:
          type: string
          enum: [post, page]
        slug:
          type: string
        title:
          type: string
...
    EditorPreviewRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
        frontMatter:
          type: object
          additionalProperties: true
    EditorPreviewResponse:
      type: object
      required: [html]
      properties:
        html:
          type: string
    SitePreview:
      type: object
      required: [status]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [Starting, Running, Failed, Stopped]
        baseUrl:
          type: string
          nullable: true
        sourceRef:
          type: string
          description: Commit SHA of saved session content used for preview
        errorMessage:
          type: string
          nullable: true
...